Support for JSX and React !!
1. I started to destructure the logic of the plugin in multiple files as we're starting to have different behavior depending on the framework, it makes things a bit more clear.

2. JSX support ! (beta)

At first, I didn't want to support React because it involved to add a jsx parser to the plugin. My opinion is that Svelte and Vue projects shouldn't download a dependency they don't need, it's not fair to make a sveltekit project download a jsx parser.

By toying with Babel I realized that I didn't need to parse JSX, but find a way to detect the XML/HTML from jsx files and perform the same operations we do on Svelte and Vue on this XML.

I tried to make a RegExp that would detect any HTML in a JSX file, but Javascript has some limitations regarding recursion in RegExp. From what I read, PHP has a great implementation of RegExp, but Javascript doesn't. There's a node package "xregexp" that helps with that, but if I need a module to use a proper RegExp in order to strip out html from jsx, it's not different from installing a jsx parser. I don't want svelte kit and vue projects to download unnecessary packages if they don't benefit it.

So I made (along with chatgpt and 3 days of work) a "parser" that uses regexp to detect nodes in a text, then some association work and string interpolation to achieve the initial goal: detecting html in jsx files without adding a new dependency. It's 140 lines of code so it makes the plugin heavier in size, but at least it's included in the plugin and doesn't pollute your node_modules folder. I checked and it doesn't slow down the process for Vue and Svelte projects, since this "parser" is not used for them.

I realized that, since it's able to scrape html from jsx, the plugin could support Solid and other JSX based frameworks. But more importantly, now that the plugin chooses the correct parser accordingly to the extension of the file it's able to parse, maybe (and I say maybe because I didn't check) it will support Astro too!

I love Astro so much, but I didn't want to support it and tell people "it supports astro, but only with astro files, svelte and vue" or something like that. I want it to support Astro to the fullest. I have to improve how the parsing works, and how the plugin determines what to parse and what he shouldn't. But it may happen.

The next step is webpack, I have to dig into their documentations, but the logic of the plugin is decoupled from vite's api, and encapsulated in a function, so it shouldn't be that hard. I don't know. I always say that and I end up taking days to achieve things.

I also want to "automatize" the tailwind configuration. If I can do that, the plugin will be so much easier to setup. Just types, and omfg I just realized I didn't work on types for React. Currently there's no way for React to know that the custom attributes are not unexpected... let's say it's in beta ^^

sorry for my bad english, usually I take time to read back what I wrote, but I don't have time for that. Have a nice day ! Whoever reads this, thank you.